--- server/backend/src/auth/createToken.ts
@@ -1,7 +1,0 @@
-import dotenv from "dotenv";
-import jwt from "jsonwebtoken";
-dotenv.config();
-const createToken = async (body: object) => {
-	return jwt.sign(body, `${process.env.TOKEN_SECRET}`, { expiresIn: '3000000s' });
-}
-export default createToken;
